pyproject.toml
[tool.poetry]
name = "nonebot_plugin_epicfree"
version = "0.2.3"
description = "EpicGameStore free games promotions plugin for NoneBot2"
authors = ["monsterxcn <monsterxcn@gmail.com>"]
documentation = "https://github.com/monsterxcn/nonebot_plugin_epicfree#readme"
license = "MIT"
homepage = "https://github.com/monsterxcn/nonebot_plugin_epicfree"
readme = "README.md"
keywords = ["nonebot", "nonebot2", "epic", "free"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
httpx = ">=0.20.0, <1.0.0"
nonebot2 = ">=2.0.0b3"
nonebot-adapter-onebot = ">=2.0.0b1"
nonebot-plugin-apscheduler = ">=0.1.0"
pytz = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"